move price.js to common/types
